+++ src/bda/plone/shop/browser/availability.py
@@ -2,6 +2,7 @@ from zope.i18nmessageid import MessageFactory
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from bda.plone.cart import (
     get_item_state,
+    get_item_data_provider,
     CartItemAvailabilityBase,
 )
 
@@ -45,10 +46,12 @@ class CartItemAvailability(CartItemAvailabilityBase):
 
     @property
     def full_available_message(self):
-        # XXX: quantity unit float
         available = self.available
         if available is None:
             available = ''
+        else:
+            if not get_item_data_provider(self.context).quantity_unit_float:
+                available = int(available)
         message = _(u'full_available_message',
                     default=u'${available} items(s) available.',
                     mapping={'available': available})
@@ -56,21 +59,24 @@ class CartItemAvailability(CartItemAvailabilityBase):
 
     @property
     def critical_available_message(self):
-        # XXX: quantity unit float
+        available = self.available
+        if not get_item_data_provider(self.context).quantity_unit_float:
+            available = int(available)
         message = _(u'critical_available_message',
                     default=u'Just ${available} items(s) left.',
-                    mapping={'available': self.available})
+                    mapping={'available': available})
         return message
 
     @property
     def overbook_available_message(self):
-        # XXX: quantity unit float
         state = get_item_state(self.context, self.request)
         overbook = self.stock.overbook
         if overbook is None:
             reservable = ''
         else:
             reservable = overbook - state.reserved
+            if not get_item_data_provider(self.context).quantity_unit_float:
+                reservable = int(reservable)
         message = _(u'overbook_available_message',
                     default=u'Item is sold out. You can pre-order '
                             u'${reservable} items. As soon as item is '